# handlers/user_handlers/user_delete.go
package user_handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/wattam/golang-auth-system/database"
	"github.com/wattam/golang-auth-system/models"
)


// swagger:operation DELETE /users/{ID} Delete
// Deletes a user by ID.
// ---
//
// parameters:
// - name: ID
//   in: path
//   description: ID of the user to delete.
//   required: true
//   type: integer
//
// responses:
//  204:
//   description: NO CONTENT
func Delete(c *gin.Context) {

	id := c.Param("id")

	database.Db.Delete(&models.User{}, id)

	c.Status(http.StatusNoContent)
}